# backend/pkg/restapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API for PoeticMetric.",
        "title": "PoeticMetric REST API",
        "termsOfService": "https://poeticmetric.com/terms-of-service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/authentication/user": {
            "get": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Read currently authenticated user.",
                "tags": [
                    "authentication"
                ],
                "summary": "Read user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/poeticmetric.AuthenticationUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responder.DetailResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Update currently authenticated user.",
                "tags": [
                    "authentication"
                ],
                "summary": "Update user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/poeticmetric.AuthenticationUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responder.DetailResponse"
                        }
                    }
                }
            }
        },
...
        "/site-reports/overview": {
            "get": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Read overview report for a site.",
                "tags": [
                    "site-reports"
                ],
                "summary": "Read overview report",
                "parameters": [
                    {
                        "type": "string",
                        "name": "browserName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "browserVersion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "countryISOCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "deviceType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "operatingSystemName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "operatingSystemVersion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "referrer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "referrerSite",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "siteID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "timeZone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmCampaign",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmContent",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmMedium",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmSource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmTerm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/poeticmetric.SiteOverviewReport"
                            }
                        }
                    }
                }
            }
        },
        "/site-reports/visitor": {
            "get": {
                "security": [
                    {
                        "UserAccessTokenAuthentication": []
                    }
                ],
                "description": "Read visitor report for a site.",
                "tags": [
                    "site-reports"
                ],
                "summary": "Read visitor report",
                "parameters": [
                    {
                        "type": "string",
                        "name": "browserName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "browserVersion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "countryISOCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "deviceType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "operatingSystemName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "operatingSystemVersion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "referrer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "referrerSite",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "siteID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "timeZone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmCampaign",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmContent",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmMedium",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmSource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "utmTerm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/poeticmetric.SiteVisitorReport"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "poeticmetric.AuthenticationUser": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isEmailVerified": {
                    "type": "boolean"
                },
                "isOrganizationOwner": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "poeticmetric.SiteOverviewReport": {
            "type": "object",
            "properties": {
                "averagePageViewDurationSeconds": {
                    "type": "integer"
                },
                "averagePageViewDurationSecondsPercentageChange": {
                    "type": "integer"
                },
                "pageViewCount": {
                    "type": "integer"
                },
                "pageViewCountPerVisitor": {
                    "type": "number"
                },
                "pageViewCountPerVisitorPercentageChange": {
                    "type": "integer"
                },
                "pageViewCountPercentageChange": {
                    "type": "integer"
                },
                "visitorCount": {
                    "type": "integer"
                },
                "visitorCountPercentageChange": {
                    "type": "integer"
                }
            }
        },
        "poeticmetric.SiteVisitorReport": {
            "type": "object",
            "properties": {
                "averageVisitorCount": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/poeticmetric.SiteVisitorReportDatum"
                    }
                },
                "intervalSeconds": {
                    "type": "integer"
                }
            }
        },
        "poeticmetric.SiteVisitorReportDatum": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "type": "string"
                },
                "visitorCount": {
                    "type": "integer"
                }
            }
        },
...
        "responder.DetailResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Human-readable detail of the response.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuthentication": {
            "type": "basic"
        },
        "UserAccessTokenAuthentication": {
            "description": "User access token authentication.",
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}